+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ftmstore-fastapi
 
+## 2.1.0 (2023-10-19)
+
+- Drop support for Python 3.10 due to dependencies
+- Implement aggregation endpoints
+- Use store resolver
+
 ## 2.0.0 (2023-09-28)
 
 - Build on top of [ftmq](https://github.com/investigativedata/ftmq/)